backend/requirements.txt
fastapi 
uvicorn 
yt-dlp 
openai-whisper
sentence-transformers 
faiss-cpu 
pydantic[dotenv] 
python-multipart 
sqlite-utils